+++ lib/src/features/database/database_repository.dart
@@ -1,5 +1,5 @@
-import 'package:coffee_shop/src/features/database/database_data_source.dart';
-import 'package:coffee_shop/src/features/database/interface_savable_data_source.dart';
+import 'package:coffee_shop/src/features/database/database/database_data_source.dart';
+import 'package:coffee_shop/src/features/database/data_source/interface_savable_data_source.dart';
 import 'package:coffee_shop/src/features/database/interface_savable_repository.dart';
 import 'package:coffee_shop/src/features/menu/data/category_data.dart';
 import 'package:coffee_shop/src/features/menu/data/product_data.dart';
